ml/requirements.txt
joblib==1.1.0
matplotlib==3.5.2
notebook==6.4.12
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scikit-surprise==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.64.0
